Module oak_remote_attestation_verification::verifier

Structs

InsecureAttestationVerifier

An instance of AttestationVerifier that succeeds iff the provided attestation is empty.

ReferenceValue

Reference values used by the verifier to appraise the attestation evidence. https://www.rfc-editor.org/rfc/rfc9334.html#name-reference-values

Traits

AttestationVerifier

A trait implementing the functionality of a verifier that appraises the attestation evidence and produces an attestation result. https://www.rfc-editor.org/rfc/rfc9334.html#name-verifier

Functions

convert_pem_to_raw

Converts a pem key to raw. Will panic if it does not like like pem.

convert_raw_to_pem

looks_like_pem

Makes a plausible guess whether the public key is in PEM format.

verify_binary_endorsement

Verifies the binary endorsement for a given measurement. More precisely:

verify_endorsement_statement

Verifies endorsement against the given reference values.

verify_endorser_public_key

Verifies that the endorser public key coincides with the one contained in the attestation.

verify_rekor_public_key

Verifies that the Rekor public key coincides with the one contained in the attestation.